+++ extensions/media2/src/main/java/com/google/android/exoplayer2/ext/media2/DefaultMediaItemConverter.java
@@ -84,7 +84,7 @@ public class DefaultMediaItemConverter implements MediaItemConverter {
 
     return new MediaItem.Builder()
         .setUri(uri)
-        .setMediaId(mediaId)
+        .setMediaId(mediaId != null ? mediaId : MediaItem.DEFAULT_MEDIA_ID)
         .setMediaMetadata(
             new com.google.android.exoplayer2.MediaMetadata.Builder().setTitle(title).build())
         .setTag(media2MediaItem)